Making some smal modifications

+++ options.go
@@ -435,7 +435,7 @@ func (opts *Options) SetMaxBytesForLevelBase(value uint64) {
 
 // SetMaxBytesForLevelMultiplier sets the max Bytes for level multiplier.
 // Default: 10
-func (opts *Options) SetMaxBytesForLevelMultiplier(value float64) {
+func (opts *Options) SetMaxBytesForLevelMultiplier(value int) {
 	C.rocksdb_options_set_max_bytes_for_level_multiplier(opts.c, C.double(value))
 }
